flaskapp/requirements.txt

Flask==2.2.0
numpy==1.26.1
pipreqs==0.4.13
requests==2.31.0
opencv-python==4.8.1.78
mediapipe==0.10.11
Werkzeug==2.2.2
template-render==0.1.0
pyrebase4==4.7.1